♻️ Improve: `client::user_record` module structure [#63]
- ✏️ Rename: to `record`
- 📦 Move: to `param` module
- ✏️ Rename: `RecordGamemode` to `Gamemode`
- ✏️ Rename: `RecordSearchCriteria` to `SearchCriteria`
